Integrated smart dashboard controls

+++ RecycleRush/src/Commands/ChassisDriveStraightForTime.h
@@ -23,13 +23,13 @@
  */
 class ChassisDriveStraightForTime: public Command {
 private:
-	float mDriveTime = 0.0;
+	double mDriveTime = 0.0;
 	std::string mKey = "";
 	bool bSmartDashCtrl = false;
-	float GetDriveTime(void);
+	double GetDriveTime(void);
 public:
 	ChassisDriveStraightForTime();
-	ChassisDriveStraightForTime(float NewTime);
+	ChassisDriveStraightForTime(double NewTime);
 	ChassisDriveStraightForTime(std::string key);
 	virtual void Initialize();
 	virtual void Execute();
